Make basic report work with parcel
