+++ .babelrc
@@ -1,6 +1,10 @@
 {
   "presets": [
-    "env",
+    ["env", {
+      "targets": {
+        "browsers": ["last 2 Chrome versions"]
+      }
+    }],
     "react"
   ],
   "plugins": ["transform-object-rest-spread"]
