+++ index.html
@@ -6,7 +6,7 @@
 </head>
 
 <body>
-  <div id="app"></div>
+  <div id="root"></div>
   <script src="src/index.jsx"></script>
 </body>
