+++ src/index.jsx
@@ -1,11 +1,22 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
+import Report from './pages/Report';
 
-class HelloMessage extends React.Component {
-  render() {
-    return <div>Hello {this.props.name}</div>;
-  }
-}
+const main = async () => {
+  const { data, calculations } = await Report.getInitialProps({
+    query: {
+      fundAddress: '0xbada55',
+      timeSpanStart: 1514761200,
+      timeSpanEnd: 1522447200,
+    },
+  });
 
-var mountNode = document.getElementById('app');
-ReactDOM.render(<HelloMessage name="Jane" />, mountNode);
+  console.log({ data, calculations });
+
+  ReactDOM.render(
+    <Report data={data} calculations={calculations} />,
+    document.getElementById('root'),
+  );
+};
+
+main();
